refactor: move lockfile schemas to `cargo-util-schemas`
